make hashes deterministic over multiple runs
We want our solver to perform deterministic. As we iterate over HashMap's, the execution depends on the order, which is different on every run as the default Hasher is initialized with a random seed. We replace the default hasher with a hasher that is always initialized with the same seed.