Improve the way universal reductions are handeled in the refinement.
There are two ways of universal variables in refinement:
* hierarichal universal variables, those get removed by the search for the next refinement level
* non-dependencies that are bound before

Before, these two cases were handeled in the same way, by adding an actual refinement clause.
Now, we do standard refinement in the first case and only add a clause in the second case.

Also, added an example which shows that the second case is unavoidable.